+++ tests/test_readme.py
@@ -2,10 +2,11 @@ from itertools import dropwhile, islice, takewhile
 from os.path import abspath, join
 from typing import Iterator
 
-from dicognito.anonymizer import Anonymizer
+import dicognito.__main__
+import pytest
 
 
-def test_anonymized_elements_described():
+def test_anonymized_elements_described(capsys):
     def before_description(line: str) -> bool:
         return not line.startswith("Using the default settings, dicognito will")
 
@@ -15,14 +16,16 @@ def test_anonymized_elements_described():
     def drop_header(iterable: Iterator[str]) -> Iterator[str]:
         return islice(iterable, 1, None)
 
+    with pytest.raises(SystemExit):
+        dicognito.__main__.main(("--what-if",))
+    anonymizer_description, _ = capsys.readouterr()
+    anonymizer_description = anonymizer_description.strip()
+
     data_root = abspath(join(__file__, "..", ".."))
     with open(join(data_root, "README.md")) as readme:
-        anonymizer = Anonymizer()
-
         readme_description = "".join(
             takewhile(still_description, drop_header(dropwhile(before_description, readme))),
         ).strip()
-        anonymizer_description = anonymizer.describe_actions()
 
     print(anonymizer_description)
     assert readme_description == anonymizer_description